+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ## Unreleased
+
+- Add `Session` protocol type. This type manages the user who signs in currently.